requirements.txt
fastapi
streamfeed-parser
uvicorn
dotenv